+++ src/modules/components/proposal/proposalVoting/proposalVotingContainer/proposalVotingContainer.tsx
@@ -1,6 +1,7 @@
 import classNames from 'classnames';
 import React, { Children, type ComponentProps } from 'react';
 import { Accordion, Card, Heading } from '../../../../../core';
+import { IProposalVotingStageProps } from '../proposalVotingStage';
 
 export interface IProposalVotingContainerProps extends ComponentProps<'div'> {
     /**
@@ -12,13 +13,17 @@ export interface IProposalVotingContainerProps extends ComponentProps<'div'> {
      */
     description: string;
     /**
-     * Active stage that will be expanded by default for multi-stage proposals.
+     * Active stage that will be expanded for multi-stage proposals.
      */
     activeStage?: string;
+    /**
+     * Callback called when the user selects a stage, to be used for expanding the current active stage for multi-stage proposals.
+     */
+    onStageClick?: (stage?: string) => void;
 }
 
 export const ProposalVotingContainer: React.FC<IProposalVotingContainerProps> = (props) => {
-    const { title, description, className, children, activeStage, ...otherProps } = props;
+    const { title, description, className, children, activeStage, onStageClick, ...otherProps } = props;
 
     const processedChildren = Children.toArray(children);
     const isMultiStage = processedChildren.length > 1;
@@ -30,9 +35,9 @@ export const ProposalVotingContainer: React.FC<IProposalVotingContainerProps> = 
                 <p className="text-base font-normal leading-normal text-neutral-500">{description}</p>
             </div>
             {isMultiStage && (
-                <Accordion.Container isMulti={false} defaultValue={activeStage}>
+                <Accordion.Container isMulti={false} value={activeStage} onValueChange={onStageClick}>
                     {processedChildren.map((child, index) =>
-                        React.isValidElement(child)
+                        React.isValidElement<IProposalVotingStageProps>(child)
                             ? React.cloneElement(child, { ...child.props, index, isMultiStage })
                             : child,
                     )}