+++ src/modules/components/proposal/proposalVoting/proposalVotingStageStatus/proposalVotingStageStatus.tsx
@@ -23,7 +23,7 @@ export interface IProposalVotingStageStatusProps extends ComponentProps<'div'> {
     /**
      * End date of the proposal in timestamp or ISO format.
      */
-    endDate: string | number;
+    endDate?: string | number;
     /**
      * Defines if the proposal is a multi-stage proposal.
      */
@@ -32,8 +32,9 @@ export interface IProposalVotingStageStatusProps extends ComponentProps<'div'> {
 
 const getStatusText = (status: ProposalVotingStatus, copy: ModulesCopy, isMultiStage?: boolean) => {
     const isSingleStagePending = !isMultiStage && status === ProposalVotingStatus.PENDING;
+    const { ACCEPTED, REJECTED, VETOED, EXPIRED } = ProposalVotingStatus;
 
-    if ([ProposalVotingStatus.ACCEPTED, ProposalVotingStatus.REJECTED].includes(status) || isSingleStagePending) {
+    if ([ACCEPTED, REJECTED, VETOED, EXPIRED].includes(status) || isSingleStagePending) {
         return copy.proposalVotingStageStatus.main.proposal;
     }
 
@@ -45,13 +46,17 @@ const statusToSecondaryText = (copy: ModulesCopy): Record<ProposalVotingStatus, 
     [ProposalVotingStatus.ACTIVE]: copy.proposalVotingStageStatus.secondary.active,
     [ProposalVotingStatus.ACCEPTED]: copy.proposalVotingStageStatus.secondary.accepted,
     [ProposalVotingStatus.REJECTED]: copy.proposalVotingStageStatus.secondary.rejected,
+    [ProposalVotingStatus.EXPIRED]: copy.proposalVotingStageStatus.secondary.expired,
     [ProposalVotingStatus.UNREACHED]: copy.proposalVotingStageStatus.secondary.unreached,
+    [ProposalVotingStatus.VETOED]: copy.proposalVotingStageStatus.secondary.vetoed,
 });
 
-const statusToIcon: Map<ProposalVotingStatus, { icon: IconType; variant: AvatarIconVariant } | undefined> = new Map([
+const statusToIcon = new Map<ProposalVotingStatus, { icon: IconType; variant: AvatarIconVariant } | undefined>([
     [ProposalVotingStatus.ACCEPTED, { icon: IconType.CHECKMARK, variant: 'success' }],
     [ProposalVotingStatus.REJECTED, { icon: IconType.CLOSE, variant: 'critical' }],
     [ProposalVotingStatus.UNREACHED, { icon: IconType.CLOSE, variant: 'neutral' }],
+    [ProposalVotingStatus.EXPIRED, { icon: IconType.CLOSE, variant: 'critical' }],
+    [ProposalVotingStatus.VETOED, { icon: IconType.CLOSE, variant: 'critical' }],
 ]);
 
 export const ProposalVotingStageStatus: React.FC<IProposalVotingStageStatusProps> = (props) => {
@@ -68,7 +73,9 @@ export const ProposalVotingStageStatus: React.FC<IProposalVotingStageStatusProps
             <div className="flex flex-row gap-0.5">
                 {status === ProposalVotingStatus.ACTIVE && (
                     <span className="text-primary-400">
-                        <Rerender>{() => formatterUtils.formatDate(endDate, { format: DateFormat.DURATION })}</Rerender>
+                        <Rerender>
+                            {() => formatterUtils.formatDate(endDate, { format: DateFormat.DURATION }) ?? '-'}
+                        </Rerender>
                     </span>
                 )}
                 {status !== ProposalVotingStatus.ACTIVE && <span className="text-neutral-800">{mainText}</span>}
@@ -78,6 +85,12 @@ export const ProposalVotingStageStatus: React.FC<IProposalVotingStageStatusProps
                 )}
                 {status === ProposalVotingStatus.REJECTED && (
                     <span className="text-critical-800">{copy.proposalVotingStageStatus.status.rejected}</span>
+                )}
+                {status === ProposalVotingStatus.VETOED && (
+                    <span className="text-critical-800">{copy.proposalVotingStageStatus.status.vetoed}</span>
+                )}
+                {status === ProposalVotingStatus.EXPIRED && (
+                    <span className="text-critical-800">{copy.proposalVotingStageStatus.status.expired}</span>
                 )}
             </div>
             {status === ProposalVotingStatus.PENDING && <Spinner size="md" variant="neutral" />}